Implements basic Hex board backend.
Created a Board class for representing the Hex board, Board will hold a
reference to a Graph and should correctly initialize the Graph so that
it's neighbor aware.

Move Node to a separate class. Change Node to know about its neighbors
instead of the graph. This should make things bit easier. Let's see.

Add a few basic tests for Node. And also a some tests for Board, ie.
testing that the graphs neighbors looks correct.